Moved SRC and Test folders to root directory.
Refactored static ReadLine class to a true singleton.
Changed unit tests to use NUnit instead of XUnit.
Fixed Ctrl+C issue #59 (Add option to return CTRL+C)
Updated version.
