+++ ReadLine.Demo/ReadLine.Demo.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <DebugType>portable</DebugType>
+    <AssemblyName>ReadLine.Demo</AssemblyName>
+    <OutputType>Exe</OutputType>
+    <PackageId>ReadLine.Demo</PackageId>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../ReadLine/ReadLine.csproj" />
+  </ItemGroup>
+
+</Project>
