+++ ReadLine.Tests/Abstractions/Console2.cs
