+++ ReadLine.Tests/AutoCompleteHandler.cs
