+++ ReadLine.Tests/KeyHandlerTests.cs
@@ -1,23 +1,25 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using NUnit.Framework;
 using ReadLine.Abstractions;
 using ReadLine.Tests.Abstractions;
-using Xunit;
 using static ReadLine.Tests.ConsoleKeyInfoExtensions;
 
 
 namespace ReadLine.Tests
 {
+    [TestFixture]
     public class KeyHandlerTests
     {
         private KeyHandler _keyHandler;
-        private readonly List<string> _history;
-        private readonly AutoCompleteHandler _autoCompleteHandler;
-        private readonly string[] _completions;
-        private readonly IConsole _console;
+        private List<string> _history;
+        private AutoCompleteHandler _autoCompleteHandler;
+        private string[] _completions;
+        private IConsole _console;
 
-        public KeyHandlerTests()
+        [SetUp]
+        public void Initialize()
         {
             _autoCompleteHandler = new AutoCompleteHandler();
             _completions = _autoCompleteHandler.GetSuggestions("", 0);
@@ -31,59 +33,65 @@ namespace ReadLine.Tests
                     .ForEach(_keyHandler.Handle);
         }
 
-        [Fact]
+        [Test]
         public void TestWriteChar()
         {
-            Assert.Equal("Hello", _keyHandler.Text);
+            Assert.AreEqual("Hello", _keyHandler.Text);
             
             " World".Select(c => c.ToConsoleKeyInfo())
                     .ToList()
                     .ForEach(_keyHandler.Handle);
                     
-            Assert.Equal("Hello World", _keyHandler.Text);
+            Assert.AreEqual("Hello World", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestBackspace()
         {
             _keyHandler.Handle(Backspace);
-            Assert.Equal("Hell", _keyHandler.Text);
+            Assert.AreEqual("Hell", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestDelete()
         {
             new List<ConsoleKeyInfo>() { LeftArrow, Delete }
                 .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("Hell", _keyHandler.Text);
+            Assert.AreEqual("Hell", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestDelete_EndOfLine()
         {
             _keyHandler.Handle(Delete);
-            Assert.Equal("Hello", _keyHandler.Text);
+            Assert.AreEqual("Hello", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
+        public void TestControlC() {
+          //_keyHandler.Handle(CtrlC);
+          Assert.Ignore("Unable to test.");
+        }
+
+        [Test]
         public void TestControlH()
         {
             _keyHandler.Handle(CtrlH);
-            Assert.Equal("Hell", _keyHandler.Text);
+            Assert.AreEqual("Hell", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestControlT()
         {
             var initialCursorCol = _console.CursorLeft;
             _keyHandler.Handle(CtrlT);
 
-            Assert.Equal("Helol", _keyHandler.Text);
-            Assert.Equal(initialCursorCol, _console.CursorLeft);
+            Assert.AreEqual("Helol", _keyHandler.Text);
+            Assert.AreEqual(initialCursorCol, _console.CursorLeft);
         }
 
-        [Fact]
+        [Test]
         public void TestControlT_LeftOnce_CursorMovesToEnd()
         {
             var initialCursorCol = _console.CursorLeft;
@@ -91,11 +99,11 @@ namespace ReadLine.Tests
             new List<ConsoleKeyInfo>() { LeftArrow, CtrlT }
                 .ForEach(_keyHandler.Handle);
             
-            Assert.Equal("Helol", _keyHandler.Text);
-            Assert.Equal(initialCursorCol, _console.CursorLeft);
+            Assert.AreEqual("Helol", _keyHandler.Text);
+            Assert.AreEqual(initialCursorCol, _console.CursorLeft);
         }
 
-        [Fact]
+        [Test]
         public void TestControlT_CursorInMiddleOfLine()
         {
             Enumerable
@@ -107,11 +115,11 @@ namespace ReadLine.Tests
 
             _keyHandler.Handle(CtrlT);
 
-            Assert.Equal("Hlelo", _keyHandler.Text);
-            Assert.Equal(initialCursorCol + 1, _console.CursorLeft);
+            Assert.AreEqual("Hlelo", _keyHandler.Text);
+            Assert.AreEqual(initialCursorCol + 1, _console.CursorLeft);
         }
 
-        [Fact]
+        [Test]
         public void TestControlT_CursorAtBeginningOfLine_HasNoEffect()
         {
             _keyHandler.Handle(CtrlA);
@@ -120,47 +128,47 @@ namespace ReadLine.Tests
 
             _keyHandler.Handle(CtrlT);
 
-            Assert.Equal("Hello", _keyHandler.Text);
-            Assert.Equal(initialCursorCol, _console.CursorLeft);
+            Assert.AreEqual("Hello", _keyHandler.Text);
+            Assert.AreEqual(initialCursorCol, _console.CursorLeft);
         }
 
-        [Fact]
+        [Test]
         public void TestHome()
         {
             new List<ConsoleKeyInfo>() { Home, 'S'.ToConsoleKeyInfo() }
                 .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("SHello", _keyHandler.Text);
+            Assert.AreEqual("SHello", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestControlA()
         {
             new List<ConsoleKeyInfo>() { CtrlA, 'S'.ToConsoleKeyInfo() }
                 .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("SHello", _keyHandler.Text);
+            Assert.AreEqual("SHello", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestEnd()
         {
             new List<ConsoleKeyInfo>() { Home, End, ExclamationPoint }
                 .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("Hello!", _keyHandler.Text);
+            Assert.AreEqual("Hello!", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestControlE()
         {
             new List<ConsoleKeyInfo>() { CtrlA, CtrlE, ExclamationPoint }
                 .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("Hello!", _keyHandler.Text);
+            Assert.AreEqual("Hello!", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestLeftArrow()
         {
             " N".Select(c => c.ToConsoleKeyInfo())
@@ -168,10 +176,10 @@ namespace ReadLine.Tests
                 .ToList()
                 .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("Hell No", _keyHandler.Text);
+            Assert.AreEqual("Hell No", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestControlB()
         {
             " N".Select(c => c.ToConsoleKeyInfo())
@@ -179,19 +187,19 @@ namespace ReadLine.Tests
                 .ToList()
                 .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("Hell No", _keyHandler.Text);
+            Assert.AreEqual("Hell No", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestRightArrow()
         {
             new List<ConsoleKeyInfo>() { LeftArrow, RightArrow, ExclamationPoint }
                 .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("Hello!", _keyHandler.Text);
+            Assert.AreEqual("Hello!", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestControlD()
         {
             Enumerable.Repeat(LeftArrow, 4)
@@ -199,44 +207,44 @@ namespace ReadLine.Tests
                     .ToList()
                     .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("Hllo", _keyHandler.Text);
+            Assert.AreEqual("Hllo", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestControlF()
         {
             new List<ConsoleKeyInfo>() { LeftArrow, CtrlF, ExclamationPoint }
                 .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("Hello!", _keyHandler.Text);
+            Assert.AreEqual("Hello!", _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestControlL()
         {
             _keyHandler.Handle(CtrlL);
-            Assert.Equal(string.Empty, _keyHandler.Text);
+            Assert.AreEqual(string.Empty, _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestUpArrow()
         {
             _history.AsEnumerable().Reverse().ToList().ForEach((history) => {
                 _keyHandler.Handle(UpArrow);
-                Assert.Equal(history, _keyHandler.Text);
+                Assert.AreEqual(history, _keyHandler.Text);
             });
         }
 
-        [Fact]
+        [Test]
         public void TestControlP()
         {
             _history.AsEnumerable().Reverse().ToList().ForEach((history) => {
                 _keyHandler.Handle(CtrlP);
-                Assert.Equal(history, _keyHandler.Text);
+                Assert.AreEqual(history, _keyHandler.Text);
             });
         }
 
-        [Fact]
+        [Test]
         public void TestDownArrow()
         {
             Enumerable.Repeat(UpArrow, _history.Count)
@@ -244,12 +252,12 @@ namespace ReadLine.Tests
                     .ForEach(_keyHandler.Handle);
 
             _history.ForEach( history => {
-                Assert.Equal(history, _keyHandler.Text);
+                Assert.AreEqual(history, _keyHandler.Text);
                 _keyHandler.Handle(DownArrow);
             });
         }
 
-        [Fact]
+        [Test]
         public void TestControlN()
         {
             Enumerable.Repeat(UpArrow, _history.Count)
@@ -257,40 +265,40 @@ namespace ReadLine.Tests
                     .ForEach(_keyHandler.Handle);
 
             _history.ForEach( history => {
-                Assert.Equal(history, _keyHandler.Text);
+                Assert.AreEqual(history, _keyHandler.Text);
                 _keyHandler.Handle(CtrlN);
             });
         }
 
-        [Fact]
+        [Test]
         public void TestControlU()
         {
             _keyHandler.Handle(LeftArrow);
             _keyHandler.Handle(CtrlU);
 
-            Assert.Equal("o", _keyHandler.Text);
+            Assert.AreEqual("o", _keyHandler.Text);
 
             _keyHandler.Handle(End);
             _keyHandler.Handle(CtrlU);
 
-            Assert.Equal(string.Empty, _keyHandler.Text);
+            Assert.AreEqual(string.Empty, _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestControlK()
         {
             _keyHandler.Handle(LeftArrow);
             _keyHandler.Handle(CtrlK);
 
-            Assert.Equal("Hell", _keyHandler.Text);
+            Assert.AreEqual("Hell", _keyHandler.Text);
 
             _keyHandler.Handle(Home);
             _keyHandler.Handle(CtrlK);
 
-            Assert.Equal(string.Empty, _keyHandler.Text);
+            Assert.AreEqual(string.Empty, _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestControlW()
         {
             " World".Select(c => c.ToConsoleKeyInfo())
@@ -298,20 +306,20 @@ namespace ReadLine.Tests
                     .ToList()
                     .ForEach(_keyHandler.Handle);
 
-            Assert.Equal("Hello ", _keyHandler.Text);
+            Assert.AreEqual("Hello ", _keyHandler.Text);
 
             _keyHandler.Handle(Backspace);
             _keyHandler.Handle(CtrlW);
 
-            Assert.Equal(string.Empty, _keyHandler.Text);
+            Assert.AreEqual(string.Empty, _keyHandler.Text);
         }
 
-        [Fact]
+        [Test]
         public void TestTab()
         {
             _keyHandler.Handle(Tab);
             // Nothing happens when no auto complete handler is set
-            Assert.Equal("Hello", _keyHandler.Text);
+            Assert.AreEqual("Hello", _keyHandler.Text);
 
             _keyHandler = new KeyHandler(new Console2(), _history, _autoCompleteHandler);
 
@@ -319,17 +327,17 @@ namespace ReadLine.Tests
 
             _completions.ToList().ForEach(completion => {
                 _keyHandler.Handle(Tab);
-                Assert.Equal($"Hi {completion}", _keyHandler.Text);
+                Assert.AreEqual($"Hi {completion}", _keyHandler.Text);
             });
         }
 
-        [Fact]
+        [Test]
         public void TestBackwardsTab()
         {
             _keyHandler.Handle(Tab);
 
             // Nothing happens when no auto complete handler is set
-            Assert.Equal("Hello", _keyHandler.Text);
+            Assert.AreEqual("Hello", _keyHandler.Text);
 
             _keyHandler = new KeyHandler(new Console2(), _history, _autoCompleteHandler);
 
@@ -340,17 +348,17 @@ namespace ReadLine.Tests
 
             _completions.Reverse().ToList().ForEach(completion => {
                 _keyHandler.Handle(ShiftTab);
-                Assert.Equal($"Hi {completion}", _keyHandler.Text);
+                Assert.AreEqual($"Hi {completion}", _keyHandler.Text);
             });
         }
 
-        [Fact]
+        [Test]
         public void MoveCursorThenPreviousHistory()
         {
             _keyHandler.Handle(LeftArrow);
             _keyHandler.Handle(UpArrow);
 
-            Assert.Equal("clear", _keyHandler.Text);
+            Assert.AreEqual("clear", _keyHandler.Text);
         }
     }
 }
