+++ ReadLine/Abstractions/Console2.cs
