+++ ReadLine/Abstractions/IConsole.cs
