+++ ReadLine/IAutoCompleteHandler.cs
