--- src/ReadLine.Demo/Program.cs
@@ -1,31 +1,0 @@
-﻿using System;
-
-
-namespace ReadLine.Demo
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            Console.WriteLine("ReadLine Library Demo");
-            Console.WriteLine("---------------------");
-            Console.WriteLine();
-
-            string[] history =
-            {
-                "ls -a",
-                "dotnet run",
-                "git init"
-            };
-            ReadLine.AddHistory(history);
-
-            ReadLine.AutoCompletionHandler = new AutoCompletionHandler();
-
-            var input = ReadLine.Read("(prompt)> ");
-            Console.WriteLine(input);
-
-            input = ReadLine.ReadPassword("Enter Password> ");
-            Console.WriteLine(input);
-        }
-    }
-}
