--- src/ReadLine.Demo/ReadLine.Demo.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <DebugType>portable</DebugType>
-    <AssemblyName>ReadLine.Demo</AssemblyName>
-    <OutputType>Exe</OutputType>
-    <PackageId>ReadLine.Demo</PackageId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="../ReadLine/ReadLine.csproj" />
-  </ItemGroup>
-
-</Project>
